requirements.txt
# ── Core Framework ──────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database & ORM ───────────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0

# ── Data & ML ────────────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.13.0

# ── HTTP & Scraping ──────────────────────────────────────────────────────────
requests==2.31.0
httpx==0.27.0
playwright==1.44.0
fake-useragent==1.5.1
beautifulsoup4==4.12.3
lxml==5.2.1

# ── Scheduling ───────────────────────────────────────────────────────────────
apscheduler==3.10.4

# ── Caching ──────────────────────────────────────────────────────────────────
redis==5.0.4
python-dotenv==1.0.1

# ── Auth / Security ──────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── AI / LangChain ───────────────────────────────────────────────────────────
langchain==0.2.1
langchain-community==0.2.1
langchain-google-genai==1.0.3
openai==1.30.1                # used for DeepSeek & OpenAI-compatible endpoints
anthropic==0.26.0

# ── Templating / Reporting ───────────────────────────────────────────────────
jinja2==3.1.4
weasyprint==62.1               # PDF generation (pure-python, no wkhtmltopdf)
markdown==3.6

# ── Notifications ────────────────────────────────────────────────────────────
python-telegram-bot==21.2

# ── WebSockets ───────────────────────────────────────────────────────────────
websockets==12.0

# ── Utilities ────────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
loguru==0.7.2
tenacity==8.3.0               # retries
pytz==2024.1
python-dateutil==2.9.0.post0
tabulate==0.9.0
rich==13.7.1                  # pretty terminal output

# ── Testing ──────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0                  # also used for TestClient
